Fixing the delete method of UserService, to throw exceptions correctly

+++ src/main/java/com/alissonfgc/jpacourse/services/UserService.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.alissonfgc.jpacourse.entities.User;
@@ -36,9 +35,9 @@ public class UserService {
 
 	public void delete(Long id) {
 		try {
-			repository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
+			if (findById(id) != null) {
+				repository.deleteById(id);
+			}
 		} catch (DataIntegrityViolationException e) {
 			throw new DatabaseException(e.getMessage());
 		}
